Former-commit-id: 3a1ecb6b7f68abe468a90ccf0cbc20ff89555d9d

+++ src/modelchecker/DFTAnalyser.h
@@ -122,7 +122,7 @@ private:
                 STORM_LOG_TRACE("Result for permutation:"<<permResult);
                 permutation = nextBitPermutation(permutation);
                 result += permResult;
-            } while(permutation < (1 << nrM) && permutation != 0);
+            } while(permutation < (static_cast<size_t>(1) << nrM) && permutation != 0);
         }
         if(invResults) {
             return storm::utility::one<ValueType>() - result;
